+++ doc/Changelog.md
@@ -1,3 +1,4 @@
+* unrecognized file are now reported with their mime type
 * when an album has many media/positions, they are saved to a separate json file
   * new option: `max_media_in_json_file`
   * new option: `max_media_from_positions_in_json_file`